wip: update example
