+++ app/build.gradle
@@ -43,14 +43,14 @@ dependencies {
     implementation 'androidx.navigation:navigation-fragment:2.5.3'
     implementation 'androidx.navigation:navigation-ui:2.5.3'
     implementation 'androidx.annotation:annotation:1.2.0'
-    implementation files('libs/tkey-native/tkey-debug.aar')
 
     implementation 'com.github.grvgoel81:customauth-android-sdk:5.0.3'
     implementation 'com.github.grvgoel81:fetch-node-details-java:4.0.0'
     implementation 'com.github.grvgoel81:torus-utils-java:5.0.0'
-    implementation 'com.github.grvgoel81:tss-client-android:0.0.8'
-    //Web3j
+    implementation 'com.github.tkey:web3-android-mpc-provider:0.0.2'
+    implementation 'com.github.tkey:tkey-android:v2-SNAPSHOT'
     implementation 'org.web3j:core:4.9.5'
+
     implementation 'com.google.code.gson:gson:2.9.1'
 
     testImplementation 'junit:junit:4.13.2'
