+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.tkey_android.FirstFragment"
+        android:name="com.example.tkey_android.TkeyMpcFragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first">
